dashboard/requirements.txt
flask==2.0.1
kafka-python==2.0.2